vol_RAM_Image.bat
Place this batch file on the desktop.  Must have volatility-2.4.standalon.exe on the desktop as well.  Take the RAM image, drag and drop it on to this batch file.  The batch file will run and execute volatility.  The output will be placed on the desktop as a text file.