Create vol_bat.bat
batch file that provides an organized output file of several volatility commands.  It works best if you have the volatility.exe program for windows on the desktop, along with the memory image and this batch file.  Then drag and drop the batch file on to the memory image file to start the script.